refactored & consolidated the util functions dealing with batch-items
